DIS-51: feat: Create Objects and DB Tables
Create Objects and DB tables for community campaigns and milestones.

feat: Add campaigns to services

Add campaigns to services so campaigns can now be added and edited by
users with the correct
permissions.

feat: Add Milestones to services

Add Milestones to services so users with the correct permissions can add
and edit them.