DIS-100: refactor: extract launchGraph()
the launch() method must be implemented on children of Admin_Admin,
taking no parameter, and returning void.

Since its logic is the same for every usagegraph class, with the
exception of the section name, I extracted it into
UsageGraphs_UsageGraphs::launchGraph($section), leaving launch() the
responsibility to assign the section name.