DIS-188: Normalize series titles so that they match without regard to case or diacritics
Series with diacritics in the titles were getting mistakenly deleted.  Using "COLLATE utf8_unicode_ci" in the matching query turned out not to be 'accent-sensitive' the way I expected and so didn't fix the problem.  Instead normalize both the series name from the record and the series name from the database when comparing them.  This also decreases the number of duplicate series caused by inconsistent use of accents.