DIS-239: refactor: update Aspen db on succcess
Since patron email are stored in the Aspen
database, we should not treat them as transient
data (which is what loadContactInfomation
handles).

This modifies how Aspen ensures that the updated
email value is shown to the user: instead of
reloading the data from Evergreen, we update its
value in Aspen db on success, so that the new
value can be fetched from Aspen db
on page reload.