DIS-76: feat: display and save URL param setting values
Note:
None of URL Parameter settings are required. Adding in a value overwrite default behaviour. Adding in a Koha additional field overwrite both default behaviour and any value entered. If a default behaviour is not simply to assign an empty string to the parameter then the tooltip will inform the user of the default behaviour.