DIS-76: feat: build HeyCentric URL and hash
Build the payment URL, taking into account the HeyCentric Setting, including its associated HeyCentric URL Parameter Settings.

For each URL parameter, assign either the associated Koha additional field value, or the value as assigned to the relevant HeyCentric Setting under "Value to assign to this URL parameter if known", or the default, in this order.
For most URL parameters, the default is an empty string. Specific default values apply to 'val1', 'val1Desc', 'am', 'rurl', and 'email'.

Notes:
Multiline URL parameter may be assigned Koha additional field values sepcific to the library and/or a given fine.
Non-multiline URL parameters may only be assigned Koha additional field values sepcific to the library.
The default for 'rurl' should ideally not be overridden as this would remove the ability for the patron to be made aware of the outcome of their payment attempt upon being redirected.