DIS-644: fix: reset curl connection options
Prevents a PUT requests from being involuntarily
sent by curlGetPage by unsetting
CURLOPT_CUSTOMREQUEST on the CurlWrapper object.

This allows instances of CurlWrapper to be used
for more than one request when not all requests
are of the same type.

Notes:
The bug (curlSendPage sending a PUT request
instead of a GET request) was replicated locally
on the MyPrivacySettings page by updating the
consent then saving, with the issue occurring on
page reload.

It was also found to occur as part of the self reg
workflow when the newsletter consent feature was
activated, though this could not be replicated
locally servers.