DIS-983: fix: log error message
Echoing the error message causes an AJAX error to
be printed to the console instead of the actual
SMTP error message being made visible.

Log the error message instead to prevent this AJAX
error and so that it is easier to diagnose any
issues with the SMTP connection occurring
upon an email being sent.